VZ200 / VZ300. All graphics C source code.
VZ200 / VZ300 / Laser 210 / Laser 310.
All of this requires the VZ200 libraries within the Z88 development kit to compile , found at : www.z88dk.org
A lot of this is simply Z80 assembly within a C wrapper because it just seemed easier to me. 
Or perhaps it was that I just couldn't be bothered assembling asm with a Z80 assembler.
Heaps of .VZ snapshots are included. Extract the zip, and simply load into a VZ emulator. example : VZEM or MAME.
Various readme files are found everywhere with utter jibberish and dribble included.